+++ packages/platform-server/src/transfer_state.ts
@@ -7,8 +7,8 @@
  */
 
 import {DOCUMENT} from '@angular/common';
-import {APP_ID, NgModule, Provider} from '@angular/core';
-import {TransferState, ɵescapeHtml as escapeHtml} from '@angular/platform-browser';
+import {APP_ID, NgModule, Provider, ɵescapeTransferStateContent as escapeTransferStateContent} from '@angular/core';
+import {TransferState} from '@angular/platform-browser';
 
 import {BEFORE_APP_SERIALIZED} from './tokens';
 
@@ -34,7 +34,7 @@ function serializeTransferStateFactory(doc: Document, appId: string, transferSto
     const script = doc.createElement('script');
     script.id = appId + '-state';
     script.setAttribute('type', 'application/json');
-    script.textContent = escapeHtml(content);
+    script.textContent = escapeTransferStateContent(content);
 
     // It is intentional that we add the script at the very bottom. Angular CLI script tags for
     // bundles are always `type="module"`. These are deferred by default and cause the transfer
